# src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RVM API",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Customer name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Customer email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Customer password.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Customer confirmation password.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Customer role.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Customer phone.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Emaill address.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get user profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/recoverpassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recover Password, Send Reset Password Mail",
                "operationId": "recoverpassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Emaill address.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/resetpassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset the user password",
                "operationId": "resetpassword",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "passtoken",
                        "in": "query",
                        "description": "Password Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/filter/organisations": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Show organisations name/id.",
                "operationId": "filter_org_name",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/filter/volunteers/courses": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Show courses by name/id.",
                "operationId": "filter_courses",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/filter/users/institutions": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Show all institutions.",
                "operationId": "filter_institution_users",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/filter/accreditedby": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Show all Accreditors.",
                "operationId": "filter_accreditedby",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/filter/map": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Show number of resources and volunteers for each county.",
                "operationId": "filter_accreditedby",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}